+++ 3-Main-SoC-Realtek-RTL8196E/32-Kernel/files/drivers/net/ethernet/rtl8196e-eth/rtl8196e_regs.h
@@ -75,6 +75,8 @@ static inline u32 rtl8196e_readl(u32 reg)
 /* LED controller (switch ASIC) */
 #define LEDCREG        (0x0100 + SWMISC_BASE)  /* 0xBB804300 */
 #define LEDMODE_DIRECT (2 << 20)               /* Bits 21:20 = 10: direct mode */
+#define DIRECTLCR      (0x0114 + SWMISC_BASE)  /* 0xBB804314 */
+#define DIRECTLCR_DEFAULT 0x1003FFFF           /* Reset value (LEDs on) */
 
 /* VLAN / netif mapping */
 #define VCR0      (0x00 + 0x4A00 + SWCORE_BASE)